requirements.txt
# Core Framework
PyQt6>=6.4.0
PyQt6-WebEngine>=6.4.0
polars>=1.9.0
pydantic>=2.0.0

# API Integration
simple-salesforce>=1.12.0
woocommerce>=3.0.0
httpx>=0.24.0
requests>=2.31.0
aiohttp>=3.9.0

# Data & Export
openpyxl>=3.1.0
xlsxwriter>=3.1.0
pyarrow>=15.0.0
numba>=0.59.0

# Security & Authentication
keyring>=24.0.0
cryptography>=41.0.0
PyJWT>=2.8.0
authlib>=1.3.0
python-dotenv>=1.0.0

# UI Enhancements
qtawesome>=1.2.0
qdarkstyle>=3.2.0

# Visualization
plotly>=5.24.0
kaleido>=0.2.1

# Advanced Export (Optional - for enhanced features)
# python-pptx>=0.6.21  # For PowerPoint export
# reportlab>=4.0.0     # For PDF reports  
# Pillow>=10.0.0       # For image processing

# Development & Testing
pytest>=7.4.0
pytest-qt>=4.2.0